services/python-service/requirements.txt
pika
minio
pillow
numpy
